+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v1.4.0] - 2021-04-07
 
 ### Added
 
@@ -395,4 +395,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v1.2.0]: https://github.com/piraeusdatastore/piraeus-operator/compare/v1.1.0...v1.2.0
 [v1.3.0]: https://github.com/piraeusdatastore/piraeus-operator/compare/v1.2.0...v1.3.0
 [v1.3.1]: https://github.com/piraeusdatastore/piraeus-operator/compare/v1.3.0...v1.3.1
-[Unreleased]: https://github.com/piraeusdatastore/piraeus-operator/compare/v1.3.1...HEAD
+[v1.4.0]: https://github.com/piraeusdatastore/piraeus-operator/compare/v1.3.1...v1.4.0